core/src/main/java/com/soffid/iam/doc/service/DocumentServiceImpl.java

+++ core/src/main/java/com/soffid/iam/doc/service/DocumentServiceImpl.java
@@ -426,15 +426,13 @@ public class DocumentServiceImpl extends DocumentServiceBase {
 		Long l = Long.decode(reference.getId());
 		result = getDocumentEntityDao().load(l);
 		
-		if (result != null) {
-			if(!result.getHash().equals(reference.getHash()))
-			{
-				throw new DocumentBeanException("No se encontro un documento con referencia " + reference);
-			}
-			
-			getNASManager().deleteFile(result.getFsPath());
-			getDocumentEntityDao().remove(result);
+		if(result== null || !result.getHash().equals(reference.getHash()))
+		{
+			return; // Document already removed
 		}
+
+		getNASManager().deleteFile(result.getFsPath());
+		getDocumentEntityDao().remove(result);
 		
 	}
